requirements.txt
Django==2.0.3
djangorestframework==3.7.7
virtualenv==15.1.0